backend/requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
matplotlib>=3.8
joblib>=1.3
requests>=2.31
beautifulsoup4>=4.12
python-whois>=0.9.4
tldextract>=5.1
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
pytest>=8.0
